Edit not saving

+++ src/main/java/br/com/uc9sa4/sa4/service/LivroService.java
@@ -9,11 +9,11 @@ import br.com.uc9sa4.sa4.models.Livro;
 public interface LivroService {
 	public List<Livro> listarLivros();
 	
-	public void apagarLivro(Integer id);
+	public void apagarLivro(Long id);
 	
 	public Livro salvarLivro(Livro livro);
 	
-	public Livro consultarLivroId(Integer id);
+	public Livro consultarLivroId(Long id);
 	
 	public Livro atualizarLivro(Livro livro);
 }
